+++ sistemaaid/backend/models/__init__.py
@@ -1,0 +1,5 @@
+from .modelCarrera import *
+from .modelSede import *
+from .modelColegio import *
+from .modelParticipante import *
+from .modelParticipanteCarrera import *
