--- sistemaaid/backend/participantes2.csv
@@ -1,3 +1,0 @@
-﻿Nombres y Apellidos,Cedula,Genero,Telefono
-Messi,55555,M,4242556589
-Cristiano Ronaldo,4235646,M,4242556590
