+++ sistemaaid/backend/views/__init__.py
@@ -1,0 +1,3 @@
+from .participanteCarrera import *
+from .participante import *
+from backend.models import *
